feature: Resolve filtering habits
