+++ client/src/actions/types.js
@@ -7,5 +7,4 @@ export default keyMirror({
 	ADD_HABIT: null,
 	DELETE_HABIT: null,
 	UPDATE_HABIT: null,
-	FILTER_HABIT_LIST: null,
 });
