+++ client/src/components/HabitList/HabitList.jsx
@@ -1,70 +1,33 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
+import { filterHabitsSelector } from '../../reducers/habitReducer';
 import { Container, NoHabits } from './HabitList.styled';
 
 import Habit from '../Habit/Habit';
 
-// eslint-disable-next-line no-unused-vars
-const habitsTest = [
-	{
-		name: 'Test habit',
-		frequency: { period: 'Once every day' },
-		status: 'Due Tomorrow',
-		description: 'Zombie ipsum reversus ab viral inferno, nam rick grimes malum',
-		type: 'Active',
-		difficulty: 'Medium',
-		id: 90672,
-		tags: ['office', 'exercise'],
-	},
-	{
-		name: 'Test habit 1',
-		frequency: { period: 'Once every month' },
-		status: 'Due',
-		description:
-			'Zombie ipsum reversus ab viral inferno, nam rick grimes malum cerebro. De carne lumbering animata corpora quaeritis.',
-		type: 'Active',
-		difficulty: 'Medium',
-		id: 72082,
-		tags: ['exercise'],
-	},
-	{
-		name: 'Test habit 2',
-		frequency: { period: 'Once every week' },
-		status: 'Due in 2 Days',
-		description: 'Zombie ipsum reversus ab viral inferno, nam rick grimes malum',
-		type: 'Active',
-		difficulty: 'Medium',
-		id: 75361,
-		tags: ['office'],
-	},
-	{
-		name: 'Test habit 3',
-		frequency: { period: 'Twice a week' },
-		status: 'Due',
-		description:
-			'Zombie ipsum reversus ab viral inferno, nam rick grimes malum cerebro. De carne lumbering animata corpora quaeritis. Summus brains sit​​.',
-		type: 'Active',
-		difficulty: 'Medium',
-		id: 17922,
-		tags: ['web development', 'study'],
-	},
-];
-
-class HabitList extends Component {
+export default class HabitList extends Component {
 	static propTypes = {
 		habits: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
+		target: PropTypes.string,
+		criteria: PropTypes.string,
+	};
+
+	static defaultProps = {
+		target: '',
+		criteria: '',
 	};
 
 	state = {};
 
 	render() {
-		const { habits } = this.props;
+		const { habits, target, criteria } = this.props;
+		const derivedHabits = filterHabitsSelector(habits, { target, criteria });
 
 		return (
 			<Container>
 				{habits ? (
-					habits.map(habit => <Habit key={habit._id} habit={habit} />)
+					derivedHabits.map(habit => <Habit key={habit._id} habit={habit} />)
 				) : (
 					<NoHabits>
 						Nothing to see here. <br /> Add some habits.
@@ -74,5 +37,3 @@ class HabitList extends Component {
 		);
 	}
 }
-
-export default HabitList;
